+++ src/EFCore.Cosmos/Query/Internal/CosmosSqlTranslatingExpressionVisitor.cs
@@ -459,11 +459,19 @@ namespace Microsoft.EntityFrameworkCore.Cosmos.Query.Internal
 
             var translation = _methodCallTranslatorProvider.Translate(_model, sqlObject, methodCallExpression.Method, arguments);
 
-            if (translation == null
-                && (methodCallExpression.Method == _stringEqualsWithStringComparison
-                    || methodCallExpression.Method == _stringEqualsWithStringComparisonStatic))
+            if (translation == null)
             {
-                ProvideTranslationErrorDetails(CoreStrings.QueryUnableToTranslateStringEqualsWithStringComparison);
+                if (methodCallExpression.Method == _stringEqualsWithStringComparison
+                    || methodCallExpression.Method == _stringEqualsWithStringComparisonStatic)
+                {
+                    ProvideTranslationErrorDetails(CoreStrings.QueryUnableToTranslateStringEqualsWithStringComparison);
+                }
+                else
+                {
+                    ProvideTranslationErrorDetails(CoreStrings.QueryUnableToTranslateMethod(
+                        methodCallExpression.Method.Name,
+                        methodCallExpression.Method.DeclaringType?.DisplayName()));
+                }
             }
 
             return translation;
